Moved `sections.rs` from `svm-sdk-macros` to `svm-cli`.
